Renamed keys.util -> keys.in_game_decorator
